Refactor: move lammps execution tools right into the lammps file
