+++ atomistics/calculators/lammps.py
@@ -5,16 +5,19 @@ from typing import TYPE_CHECKING
 from pylammpsmpi import LammpsASELibrary
 
 from atomistics.calculators.wrapper import task_evaluation
-from atomistics.calculators.lammps_library.calculator import (
-    _run_simulation,
-    lammps_input_template as _lammps_input_template,
-)
 
 if TYPE_CHECKING:
     from ase import Atoms
     from pandas import DataFrame
     from pylammpsmpi import LammpsASELibrary
     from atomistics.calculators.interface import TaskName
+
+
+LAMMPS_INPUT_TEMPLATE = """\
+thermo_style custom step temp pe etotal pxx pxy pxz pyy pyz pzz vol
+thermo_modify format float %20.15g
+thermo 100
+run 0"""
 
 
 @task_evaluation
@@ -24,7 +27,7 @@ def evaluate_with_lammps_library(
     potential_dataframe: DataFrame,
     lmp: LammpsASELibrary,
 ):
-    lmp = _run_simulation(structure, potential_dataframe, _lammps_input_template, lmp)
+    lmp = _run_simulation(structure, potential_dataframe, LAMMPS_INPUT_TEMPLATE, lmp)
     results = {}
     if "calc_energy" in tasks:
         results["energy"] = lmp.interactive_energy_pot_getter()
@@ -57,3 +60,25 @@ def evaluate_with_lammps(
     results_dict = evaluate_with_lammps_library(task_dict, potential_dataframe, lmp)
     lmp.close()
     return results_dict
+
+
+def _run_simulation(structure, potential_dataframe, input_template, lmp):
+    # write structure to LAMMPS
+    lmp.interactive_structure_setter(
+        structure=structure,
+        units="metal",
+        dimension=3,
+        boundary=" ".join(["p" if coord else "f" for coord in structure.pbc]),
+        atom_style="atomic",
+        el_eam_lst=potential_dataframe.Species,
+        calc_md=False,
+    )
+
+    # execute calculation
+    for c in potential_dataframe.Config:
+        lmp.interactive_lib_command(c)
+
+    for l in input_template.split("\n"):
+        lmp.interactive_lib_command(l)
+
+    return lmp
