+++ atomistics/calculators/lammps_library/calculator.py
@@ -1,27 +1,3 @@
-lammps_input_template = """\
-thermo_style custom step temp pe etotal pxx pxy pxz pyy pyz pzz vol
-thermo_modify format float %20.15g
-thermo 100
-run 0"""
 
 
-def _run_simulation(structure, potential_dataframe, input_template, lmp):
-    # write structure to LAMMPS
-    lmp.interactive_structure_setter(
-        structure=structure,
-        units="metal",
-        dimension=3,
-        boundary=" ".join(["p" if coord else "f" for coord in structure.pbc]),
-        atom_style="atomic",
-        el_eam_lst=potential_dataframe.Species,
-        calc_md=False,
-    )
 
-    # execute calculation
-    for c in potential_dataframe.Config:
-        lmp.interactive_lib_command(c)
-
-    for l in input_template.split("\n"):
-        lmp.interactive_lib_command(l)
-
-    return lmp
